[#3348] Start refatoring of account management service
